fix/act_migration

+++ db/migrate/20231222161040_add_missing_taggable_index.acts_as_taggable_on_engine.rb
@@ -3,11 +3,16 @@
 
 class AddMissingTaggableIndex < ActiveRecord::Migration[6.0]
   def self.up
-    add_index ActsAsTaggableOn.taggings_table, %i[taggable_id taggable_type context],
-              name: 'taggings_taggable_context_idx'
+    unless index_exists?(ActsAsTaggableOn.taggings_table, [:taggable_id, :taggable_type, :context], name: 'taggings_taggable_context_idx')
+      add_index ActsAsTaggableOn.taggings_table, [:taggable_id, :taggable_type, :context],
+                name: 'taggings_taggable_context_idx'
+    end
   end
 
   def self.down
-    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
+    if index_exists?(ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx')
+      remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
+    end
   end
 end
+
